requirements.txt
flask
pandas
ta
python-binance
scikit-learn
plotly
requests
python-dotenv